Logger😵 Log🐣 && Moved the Date formatters from Logger to Date
Changed the logging tool from `Logger.swift` to `Log.swift`, this comes
with new tools, and more customisation, when needed more will be added.

Adds the ability to show the current thread that was been used and
allows the tester to select wether the source for the call should be
shown or not.