Finally Fixed a Trolley/Core bug 🐛 & TRLUser has updated description
Fixed a bug in the currency converter that was annoying me but forgot
about it until we have added the ability to convert currency again in
Trolley.io after the move from the old project.

Was all down to rounding up the conversion rate for the decimal
numbers. 😂 Was so simple.