Example Number 2 Created (Searching Shop) + Moved Example 1 Files Around
As with any tests and new features we hit bugs, this was more so with
Example 2, when testing it all seemed to work fine, but with not using
some networking calls before and adding new ones since testing we ran
into some small issues.

1) Filtering was not working, this was fixed in the last commit [] and
was just a simple oversight that when testing i never actually tested
for “stor” or any other half complete words when wanting “storm” or
other values so it always worked.

2) Percentages were wrong when working with money, i foolishly forgot
‘how to math’ and had my formula for the percentage wrong, so i was
getting huge discount values 😂 its been a while since A-level maths so
please don’t judge, fixing that lead me to adding a new property to
Percentage -> decimalValue which has the % as its decimal (not
rounded), this is then used in the multiplier to get the discount.

3) I didn’t have a `responseProduct()` for just a singular product
before, which i didn’t think much of, thought just using the `.product`
property from JSON would work fine but this was not the case for
getting an individual property as the JSON raw array is empty so cannot
work the same, fixed a work around but decided that instead i would
just react the `responseProduct()` instead, the workaround is still
there incase its needed later on.

Also created is the new `SearchableProduct`, a stripped down version of
Product that will only hold the name, company name and id, to allow for
faster downloads as it only has to download less data and then all that
is required for the full download is to call
`Products.getProduct(with:)` to download the full details