WebSocket Updated/Fixed && Connected to Analytics
The Basic Analytics is updated and created for use on the sever, the
devices now connect, and once the connection is satisfied then
analytics sends it basic data over, the deviceToken and device details
and if its a new user or old user for the which the sever saves, its
very basic at the moment but will set this as my main goal for now as
most everything else is sorted.

WebSocket is updated with the TRLTimer so that it now times-out better
and will attempt to reconnect, only if the Device is passing
`.isReachable`, if not it will wait until its reachable to call itself
again.